+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+- `statement::bind_reference` overloads that accept callback called on dereference
+
 ## [1.2] - 2024-09-10
 
 ### Added